Fix perfoutput_func_type
Second argument is of type i64 (result of calling llvm.bpf.pseudo), 4th argument
is a structure that is defined dynamically, so just use type of the passed
data Value.

With this change `bpftrace -e 'BEGIN { printf("%s", "a"); }'` does not crash
anymore when asserts are enabled in llvm, clang, bcc and bpftrace, but there
are more type correcness issues to be fixed.